Introduce Flask-RESTX
- reorganize the endpoints using Flask-RESTX
- update some endpoint names
- add `get_state()` fuction to redis_util
