+++ app/apis/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2020, Ko Sugawara
+# Copyright (c) 2022, Ko Sugawara
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -22,42 +22,47 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 # ==============================================================================
-"""Utils used for Redis."""
+from flask_restx import Api
 
-from enum import Enum
+from elephant import __version__
 
-import redis
-
-from elephant.util import RUN_ON_FLASK
-
-REDIS_KEY_LR = 'lr'
-REDIS_KEY_NCROPS = 'n_crops'
-REDIS_KEY_STATE = 'state'
-REDIS_KEY_TIMEPOINT = 'timepoint'
-REDIS_KEY_UPDATE_ONGOING_SEG = 'update_ongoing_seg'
-REDIS_KEY_UPDATE_ONGOING_FLOW = 'update_ongoing_flow'
+from .dataset import api as dataset
+from .download import api as download
+from .flow import api as flow
+from .params import api as params
+from .seg import api as seg
+from .state import api as state
+from .upload import api as upload
 
 
-class TrainState(Enum):
-    IDLE = 0
-    RUN = 1
-    WAIT = 2
-
-
-if RUN_ON_FLASK:
-    redis_client = redis.Redis.from_url('redis://localhost:6379/0')
-else:
-    redis_client = None
-
-
-def get_state():
+def init_api(app):
     """
-    Returns the current state.
+    Initialize Api and returns it.
+
+    Parameters
+    ----------
+    app : Flask
+        Flask object.
 
     Returns
     -------
-    state : int
-        current state
+    api : flask_restx.Api
+        Api object.
 
     """
-    return int(redis_client.get(REDIS_KEY_STATE))
+    api = Api(
+        app=app,
+        version=__version__,
+        title='ELEPHANT',
+        description='ELEPHANT API',
+    )
+
+    api.add_namespace(dataset, path='/dataset')
+    api.add_namespace(download, path='/download')
+    api.add_namespace(flow, path='/flow')
+    api.add_namespace(params, path='/params')
+    api.add_namespace(seg, path='/seg')
+    api.add_namespace(state, path='/state')
+    api.add_namespace(upload, path='/upload')
+
+    return api
