+++ docker/nginx.conf
@@ -13,7 +13,8 @@ http {
         '$status $body_bytes_sent "$http_referer" '
         '"$http_user_agent" "$http_x_forwarded_for"';
     map $request_uri $loggable {
-        /gpus 0;
+        /state/gpus 0;
+        /state/gpus/ 0;
         default 1;
     }
     access_log /var/log/nginx/access.log main if=$loggable;
