+++ docker/uwsgi.ini
@@ -14,5 +14,5 @@ need-app = true
 die-on-term = true
 ; For debugging and testing
 show-config = true
-; Do not log /gpus
-route = ^/gpus donotlog:
+; Do not log /state/gpus
+route = ^/state/gpus* donotlog:
